Edit FeeInfo To SpecifiedTradeAllowanceCharge and SpecifiedTradeAllowanceChargeTest
- Add SpecifiedTradeAllowanceCharge
- Add SpecifiedTradeAllowanceChargeTest

**edit normal form to lombok form model
